requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
langchain-experimental
langchainhub
langchain_chroma
chromadb==0.4.4
flask
pypdf
GitPython
unstructured
-e .
